save state on some small cleanup
