+++ benchmarks/run_benchmark.sh
@@ -6,8 +6,7 @@ set -u
 
 function cleanup() {
     if [[ "$OSTYPE" == "darwin"* ]]; then
-        #docker-machine stop default
-        docker rm -f "$(docker ps -a -q)" || true
+        docker rm -f $(docker ps -a -q) || true
     fi
 }
 
@@ -16,17 +15,10 @@ trap cleanup EXIT
 BAZEL="bazel"
 
 if [[ "$OSTYPE" == "darwin"* ]]; then
-    docker-machine start default || true
     # On OSX we update the docker env vars to the latest
     eval "$(docker-machine env default)"
     # We also update the output location bazel uses, to make sure
     # that we will be able to map paths.
-
-    # sometimes needs 
-    #sudo spctl --master-disable
-    #bazel info workspace
-    #sudo spctl --master-enable
-
     # TODO(oschaaf): does this work on Linux?:
     export TEST_SERVER_EXTERNAL_IP="$(docker-machine ip)"
 fi
@@ -43,7 +35,7 @@ export NH_CLIENT_PATH=nighthawk_client
 export NH_OUTPUT_TRANSFORM_PATH=nighthawk_output_transform
 export TMPDIR="$(pwd)/benchmarks/tmp"
 export NH_NH_DOCKER_IMAGE="envoyproxy/nighthawk-dev:latest"
-export ENVOY_DOCKER_IMAGE_TO_TEST="envoyproxy/envoy-dev:74290ef76a76fbbf50f072dc33438791f93f68c7"
+export ENVOY_DOCKER_IMAGE_TO_TEST="envoyproxy/envoy-dev:f61b096f6a2dd3a9c74b9a9369a6ea398dbe1f0f"
 
 # run all tests starting with test_http_h1_small
 bazel-bin/benchmarks/benchmarks --log-cli-level=info -vvvv -k test_http_h1_small benchmarks/
