+++ benchmarks/test_benchmark.py
@@ -1,26 +1,22 @@
 #!/usr/bin/env python3
+"""@package integration_test.py
 
-# NOTE: this is just a demo for now, to show how this works.
+Just a demo for now. Show how to tap into Nighthawk's
+integration test framework to run benchmark executions.
+"""
+
 
 import logging
 import json
-import os
 import pytest
-
-from test.integration.common import IpVersion
-from test.integration.integration_test_fixtures import (http_test_server_fixture,
-                                                        https_test_server_fixture,
-                                                        HttpIntegrationTestBase,
-                                                        determineIpVersionsFromEnvironment)
+from test.integration.integration_test_fixtures import http_test_server_fixture
 from test.integration.utility import *
-from test.integration.nighthawk_test_server import NighthawkTestServer
-
+from infra import *
 
 def run_with_cpu_profiler(fixture,
                           rps=999999,
-                          use_h2=False,
-                          duration=5,
-                          max_connections=1,
+                          duration=1,
+                          max_connections=100,
                           max_active_requests=1,
                           request_body_size=0,
                           response_size=10,
@@ -29,16 +25,16 @@ def run_with_cpu_profiler(fixture,
     assert (fixture.proxy_server.enableCpuProfiler())
   assert (fixture.test_server.enableCpuProfiler())
   args = [
-      fixture.getTestServerRootUri(), "--rps",
-      str(rps), "--duration",
-      str(duration), "--connections",
-      str(max_connections), "--max-active-requests",
-      str(max_active_requests), "--concurrency",
-      str(concurrency), "--request-header",
-      "x-nighthawk-test-server-config:{response_body_size:%s}" % response_size
+      fixture.getTestServerRootUri(),
+      "--rps", str(rps),
+      "--duration", str(duration),
+      "--connections", str(max_connections),
+      "--max-active-requests", str(max_active_requests),
+      "--concurrency", str(concurrency),
+      "--request-header", "x-nighthawk-test-server-config:{response_body_size:%s}" % response_size,
+      "--experimental-h1-connection-reuse-strategy", "lru",
+      "--prefetch-connections"
   ]
-  if use_h2:
-    args.append("--h2")
   if request_body_size > 0:
     args.append("--request-body-size")
     args.append(str(request_body_size))
@@ -47,7 +43,7 @@ def run_with_cpu_profiler(fixture,
   counters = fixture.getNighthawkCounterMapFromJson(parsed_json)
   response_count = counters["benchmark.http_2xx"]
   request_count = counters["upstream_rq_total"]
-  connection_counter = "upstream_cx_http2_total" if use_h2 else "upstream_cx_http1_total"
+  connection_counter = "upstream_cx_http1_total"
 
   # Some arbitrary sanity checks 
   assertCounterGreater(counters, "benchmark.http_2xx", 1000)
@@ -57,66 +53,16 @@ def run_with_cpu_profiler(fixture,
 
   # Could potentially set thresholds on acceptable latency here.
 
-  # dump output
-  logging.info(fixture.transformNighthawkJson(json.dumps(parsed_json)))
+  # dump human readably output to logs
+  logging.info(fixture.transformNighthawkJson(json.dumps(parsed_json), "human"))
 
+  # TODO(oschaaf): dump fortio/json/yaml/human output formats as artifacts
 
-class EnvoyProxyServer(NighthawkTestServer):
-  def __init__(self,
-               server_binary_path,
-               config_template_path,
-               server_ip,
-               ip_version,
-               parameters=dict(),
-               tag=""):
-    super(EnvoyProxyServer, self).__init__(server_binary_path, config_template_path, server_ip,
-                                              ip_version, parameters=parameters, tag=tag)
-    self.docker_image = os.getenv("ENVOY_DOCKER_IMAGE_TO_TEST")
-
-class InjectHttpProxyIntegrationTestBase(HttpIntegrationTestBase):
-  """
-  Base for running plain http tests against the Nighthawk test server
-  """
-
-  def __init__(self, ip_version):
-    """See base class."""
-    super(InjectHttpProxyIntegrationTestBase, self).__init__(ip_version)
-
-  def setUp(self):
-    super(InjectHttpProxyIntegrationTestBase, self).setUp()
-    logging.info("injecting envoy proxy ...")
-    # TODO(oschaaf): how should this interact with multiple backends?
-    self.parameters["proxy_ip"] = self.test_server.server_ip
-    self.parameters["server_port"] = self.test_server.server_port
-    proxy_server = EnvoyProxyServer("envoy",
-                                    "benchmarks/configurations/envoy_proxy.yaml", self.server_ip,
-                                    self.ip_version, parameters=self.parameters, tag=self.tag)
-    assert (proxy_server.start())
-    logging.info("envoy proxy listening at {ip}:{port}".format(ip=proxy_server.server_ip, port=proxy_server.server_port))
-    self.proxy_server = proxy_server
-
-  def getTestServerRootUri(self):
-    """See base class."""
-    r = super(InjectHttpProxyIntegrationTestBase, self).getTestServerRootUri()
-    # TODO(oschaaf): fix, kind of a hack.
-    r = r.replace(":%s" % self.test_server.server_port, ":%s" % self.proxy_server.server_port)
-    return r
-
-@pytest.fixture(params=determineIpVersionsFromEnvironment())
-def inject_envoy_http_proxy_fixture(request):
-  '''
-  Injects an Envoy proxy.
-  '''
-  f = InjectHttpProxyIntegrationTestBase(request.param)
-  f.setUp()
-  yield f
-  f.tearDown()
-
-
-# Plain http: baseline vs running via Envoy
+# Test via injected Envoy
 def test_http_h1_small_request_small_reply_via(inject_envoy_http_proxy_fixture):
   run_with_cpu_profiler(inject_envoy_http_proxy_fixture)
 
+# Test the origin directly, using a stock fixture 
 def test_http_h1_small_request_small_reply_direct(http_test_server_fixture):
   run_with_cpu_profiler(http_test_server_fixture)
